backend/requirements.txt
fastapi==0.95.0
timeout-decorator==0.5.0
uvicorn==0.21.1
mysql-connector==2.2.9